Revert "Revert "Merge tag 'MP_3.45.0_BCM21664Android_SystemRel_6.3.10' into sdb-linux-3.10.0-hawaii-java""
This reverts commit 51589aadb334f8b9b73a57b6277fbe6a186f3824.

Conflicts:
	arch/arm/boot/dts/java_kivu_xdot0.dts
	drivers/video/backlight/pwm_bl.c

Reviewed-on: http://mps-gerrit.broadcom.com/45769
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>

+++ drivers/char/broadcom/fuse_log/output.c
@@ -293,11 +293,10 @@ static void WriteToLogDev_SDCARD(void)
 						   nFifo,
 						   &g_devWrParms.
 						   file->f_pos);
-
 				if (nWrite > 0)
 					BCMLOG_FifoRemove(&g_fifo, nWrite);
 
-				if ((nWrite < nFifo)
+				if ((nWrite < (int)nFifo)
 				    || ((Get_SDCARD_Available()) <
 					MTT_SD_RESERVED)) {
 					nFifo = 0;
